+++ src/main/kotlin/com/pestphp/pest/configuration/PestRunConfigurationHandler.kt
@@ -2,6 +2,7 @@ package com.pestphp.pest.configuration
 
 import com.intellij.execution.ExecutionException
 import com.intellij.openapi.project.Project
+import com.jetbrains.php.composer.lib.ComposerLibraryManager
 import com.jetbrains.php.config.commandLine.PhpCommandSettings
 import com.jetbrains.php.testFramework.run.PhpTestRunConfigurationHandler
 import com.pestphp.pest.toPestTestRegex
@@ -71,9 +72,13 @@ class PestRunConfigurationHandler : PhpTestRunConfigurationHandler {
         }
         val pathMapper = phpCommandSettings.pathProcessor.createPathMapper(project)
 
+        val rootPath =
+            ComposerLibraryManager.getInstance(project).findVendorDirForUpsource()?.parent?.path
+                ?: workingDirectory
+
         phpCommandSettings.addPathArgument(file)
         phpCommandSettings.addArgument(
-            "--filter=/${methodName.toPestTestRegex(workingDirectory, file, pathMapper)}/"
+            "--filter=/${methodName.toPestTestRegex(rootPath, file, pathMapper)}/"
         )
     }
 }